Wait in ITOPOD for bossess to show
